+++ src/main/kotlin/com/asap/asapbackend/client/oauth/OAuthConfig.kt
@@ -1,7 +1,7 @@
 package com.asap.asapbackend.client.oauth
 
 import com.asap.asapbackend.client.oauth.kakao.KakaoSocialLoginHandler
-import com.asap.asapbackend.domain.user.domain.enum.Provider
+import com.asap.asapbackend.domain.user.domain.model.Provider
 import com.asap.asapbackend.domain.user.domain.service.SocialLoginHandler
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
